+++ CHANGES.rst
@@ -38,6 +38,9 @@ Changes in 1.3.6 (2021-05-07)
  * Add Jitsi widget remove banner for privileged users.
  * Update "Jump to unread" banner to a pill style button.
  * CallVC: Add transfer button.
+ * Spaces: Hide spaces from room list and home but keep space invites (#4252).
+ * Spaces: Show space invites and advertise that they are not available (#4277).
+ * Advertise that spaces are not available when tapping on a space link or a space invite (#4279).
 
 🐛 Bugfix
  * RoomVC: Avoid navigation to integration management using integration popup with settings set to integration disabled (#4261).